requirements.txt
django
djangorestframework
drf_spectacular
requests
psycopg2-binary
python-dotenv